A lot of modifications and integration tests for UserController, WorkoutController and ExerciseController

+++ src/main/java/com/fontys/fitquest/business/implementation/DeleteExerciseUseCaseImpl.java
@@ -1,6 +1,7 @@
 package com.fontys.fitquest.business.implementation;
 
 import com.fontys.fitquest.business.DeleteExerciseUseCase;
+import com.fontys.fitquest.business.exception.ExerciseNotFoundException;
 import com.fontys.fitquest.persistence.ExerciseRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,9 @@ public class DeleteExerciseUseCaseImpl implements DeleteExerciseUseCase {
 
     @Override
     public void deleteExercise(int id) {
+        if (!exerciseRepository.existsById(id)) {
+            throw new ExerciseNotFoundException("Exercise not found with id: " + id);
+        }
         exerciseRepository.deleteById(id);
     }
 }
